+++ plone/recipe/codeanalysis/chameleonlint.py
@@ -5,6 +5,7 @@ from plone.recipe.codeanalysis.analyser import Analyser
 from plone.recipe.codeanalysis.analyser import console_factory
 
 import io
+import os
 import re
 import sys
 
@@ -32,33 +33,48 @@ class ChameleonLint(Analyser):
         # Please the ABC by faux-implementing the cmd.
         pass
 
-    def run(self):
+    def get_files(self):
         files = []
         for extension in self.extensions:
             files.extend(self.find_files('.*\.{0}'.format(extension)))
+        return files
 
+    def lint(self, file_content, file_path):
+        offset = 0
+        if '<!DOCTYPE' not in file_content:
+            file_content = DOCTYPE_WRAPPER.format(file_content)
+            offset = len(DOCTYPE_WRAPPER.splitlines()) - 1
+        try:
+            parse(io.StringIO(unicode(file_content)))
+        except XMLSyntaxError as e:
+            # Line number offset correction.
+            msg = e.msg
+            for line_number in re.findall('line ([0-9]+)', msg):
+                msg = msg.replace(
+                    'line {0}'.format(line_number),
+                    'line {0}'.format(int(line_number) - offset))
+            return ['{0}: {1}'.format(file_path, msg)]
+        return None
+
+    def linter(self, files):
         total_errors = []
+        files = sorted(files, key=lambda f: (len(f.split(os.sep)), f))
         for file_path in files:
             file_content = open(file_path, 'r').read()
-            offset = 0
-            if '<!DOCTYPE' not in file_content:
-                file_content = DOCTYPE_WRAPPER.format(file_content)
-                offset = len(DOCTYPE_WRAPPER.splitlines()) - 1
-            try:
-                parse(io.StringIO(unicode(file_content)))
-            except XMLSyntaxError as e:
-                # Line number offset correction.
-                msg = e.msg
-                for line_number in re.findall('line ([0-9]+)', msg):
-                    msg = msg.replace(
-                        'line {0}'.format(line_number),
-                        'line {0}'.format(int(line_number) - offset))
+            errors = self.lint(file_content, file_path)
+            if errors is not None:
+                total_errors.extend(errors)
+        return total_errors
 
-                total_errors.append('{0}: {1}'.format(file_path, msg))
-
+    def report(self, errors):
         with self.open_output_file() as output_file:
-            output_file.write('\n'.join(total_errors))
-            return self.parse_output(output_file, bool(total_errors))
+            output_file.write('\n'.join(errors))
+            return self.parse_output(output_file, bool(errors))
+
+    def run(self):
+        files = self.get_files()
+        errors = self.linter(files)
+        return self.report(errors)
 
 
 def console_script(options):